+++ src/main/java/com/umc/hwaroak/event/listener/FriendEventListener.java
@@ -1,41 +1,43 @@
-package com.umc.hwaroak.listener;
+package com.umc.hwaroak.event.listener;
 
 import com.umc.hwaroak.converter.AlarmConverter;
 import com.umc.hwaroak.domain.Alarm;
 import com.umc.hwaroak.domain.common.AlarmType;
-import com.umc.hwaroak.event.FireSendEvent;
+import com.umc.hwaroak.event.FriendRequestEvent;
 import com.umc.hwaroak.infrastructure.publisher.RedisPublisher;
 import com.umc.hwaroak.infrastructure.transaction.CustomTransactionSynchronization;
 import com.umc.hwaroak.repository.AlarmRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 
-@Component
 @RequiredArgsConstructor
-public class FireEventListener{
+@Component
+public class FriendEventListener {
 
     private final AlarmRepository alarmRepository;
     private final RedisPublisher redisPublisher;
 
     /**
-     *  불씨 보냈을시 알람 생성하기
+     *  친구 요청시 알람 생성하기
      */
-    @EventListener
-    @Transactional
-    public void sendFireAlarm(FireSendEvent event) {
-        String nickname = event.getSender().getNickname();
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
+    public void sendFriendRequestAlarm(FriendRequestEvent event) {
 
         Alarm alarm = Alarm.builder()
                 .sender(event.getSender())
                 .receiver(event.getReceiver())
-                .alarmType(AlarmType.FIRE)
-                .title("불 키우기")
-                .message(nickname + "님께서 불씨를 지폈어요!")
-                .content(nickname + "님께서 불씨를 지폈어요!")
+                .alarmType(AlarmType.FRIEND_REQUEST)
+                .title("친구 요청")
+                .message(event.getSender().getNickname() + "님이 친구 요청을 보냈습니다.")
+                .content(event.getSender().getNickname() + "님이 친구 요청을 보냈습니다.")
                 .build();
+
         alarmRepository.save(alarm);
         TransactionSynchronizationManager.registerSynchronization(
                 new CustomTransactionSynchronization() {
