+++ src/main/java/com/umc/hwaroak/event/listener/ItemEventListener.java
@@ -1,4 +1,4 @@
-package com.umc.hwaroak.listener;
+package com.umc.hwaroak.event.listener;
 
 import com.umc.hwaroak.domain.Item;
 import com.umc.hwaroak.domain.Member;
@@ -12,9 +12,11 @@ import com.umc.hwaroak.repository.MemberRepository;
 import com.umc.hwaroak.response.ErrorCode;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.event.TransactionPhase;
+import org.springframework.transaction.event.TransactionalEventListener;
 
 import java.util.List;
 import java.util.Optional;
@@ -29,8 +31,8 @@ public class ItemEventListener {
     private final MemberItemRepository memberItemRepository;
 
     // 수령 가능 아이템 추가하기
-    @EventListener
-    @Transactional
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void upgradeNextItem(ItemUpdateEvent event) {
 
         Member member = memberRepository.findById(event.getMemberId())
@@ -56,8 +58,8 @@ public class ItemEventListener {
     }
 
     // 삭제 시 이전으로 돌아가기
-    @EventListener
-    @Transactional
+    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void backToStatus(ItemRollbackEvent event) {
 
         Member member = memberRepository.findById(event.getMemberId())
